+++ src/model/printable.rs
@@ -179,11 +179,18 @@ impl Printable for ArbitraryData {
     }
 }
 
+impl Printable for Vec<u8> {
+    fn to_printable(&self, style: PrintableStyle) -> String {
+        String::from_utf8_lossy(self).to_printable(style)
+    }
+}
+
 #[cfg(test)]
 mod tests {
+    use base64_simd::STANDARD as base64;
+
     use crate::model::{Printable, PrintableStyle};
     use crate::rg::de::ArbitraryData;
-    use base64_simd::STANDARD as base64;
 
     const NON_PRINTABLE_WHITESPACE: &str = "\x00\x01\x02\x03\x04\x05\x06\x07\x08\x09\x0A\x0B\x0C\x0D\x0E\x0F\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1A\x1B\x1C\x1D\x1E\x1F\x20\x7F";