+++ src/replace.rs
@@ -1,5 +1,4 @@
-use std::fs::File;
-use std::fs::OpenOptions;
+use std::fs::{File, OpenOptions};
 use std::io::{Read, Write};
 
 use anyhow::{anyhow, Context, Result};
@@ -68,6 +67,8 @@ fn perform_replacements_in_file(
         let offset = item.offset().unwrap();
         log::debug!("Item[{}] offset: {}", i, offset);
 
+        let mut byte_buf = Vec::new();
+
         // Iterate backwards so the offset doesn't change as we make replacements.
         for (i, sub_item) in item
             .sub_items()
@@ -84,19 +85,45 @@ fn perform_replacements_in_file(
             let matched_bytes = text.to_vec();
 
             if str_to_remove.as_bytes() == matched_bytes.as_slice() {
+                // compute replacement
+                let replacement = match criteria
+                    .capture_pattern
+                    .as_ref()
+                    .and_then(|re| re.captures(&matched_bytes))
+                {
+                    // user passed a capturing group
+                    Some(captures) => {
+                        // empty buf without changing capacity
+                        byte_buf.clear();
+                        captures.expand(&criteria.user_replacement, &mut byte_buf);
+                        byte_buf.as_slice()
+                    }
+                    // just use raw replacement
+                    None => criteria.user_replacement.as_slice(),
+                };
+
+                // have to save this because it will be invalid after the replacement
                 let removed_str = str_to_remove.to_string();
-                file_as_str.replace_range(normalised_range, &criteria.text);
+                // must convert to strings since due to encoding support we perform replacements as strings
+                let replacement = std::str::from_utf8(&replacement)?;
+                // performance replacement
+                file_as_str.replace_range(normalised_range, replacement);
 
                 log::debug!(
                     "Replacement - reported line: {:?}, removed: \"{}\", added: \"{}\"",
                     item.line_number(),
                     removed_str,
-                    criteria.text
+                    replacement
                 );
             } else {
                 log::warn!("Matched bytes do not match bytes to replace!");
                 log::warn!("\tFile: \"{}\"", path_buf.display());
                 log::warn!("\tMatch: data=\"{}\", bytes={:?}", text, matched_bytes);
+                log::warn!(
+                    "\tBytes: data=\"{}\", bytes={:?}",
+                    str_to_remove,
+                    str_to_remove.as_bytes()
+                );
                 log::warn!("\tOffset: {}", offset + range.start);
                 did_skip_replacement = true;
             }
@@ -151,13 +178,18 @@ fn perform_replacements_in_file(
 
 pub fn perform_replacements(criteria: ReplacementCriteria) -> Result<()> {
     log::trace!("--- PERFORM REPLACEMENTS ---");
-    log::debug!("Replacement text: \"{}\"", criteria.text);
+    log::debug!(
+        "Replacement text: \"{}\"",
+        String::from_utf8_lossy(&criteria.user_replacement)
+    );
 
     let rg_encoding = RgEncoding::from(&criteria.encoding);
     log::debug!("User passed encoding: {:?}", rg_encoding);
 
     // Group items by their file so we only open each file once.
     let mut did_skip_replacement = false;
+
+    // TODO: consider concurrent replacements here - make it configurable - we don't want to read in multiple large files at once
     for meta in criteria.as_map() {
         match perform_replacements_in_file(&criteria, &rg_encoding, meta) {
             Ok(did_skip) => {
@@ -190,6 +222,7 @@ mod tests {
 
     use base64_simd::STANDARD as base64;
     use pretty_assertions::assert_eq;
+    use regex::bytes::Regex;
     use tempfile::NamedTempFile;
 
     use crate::model::*;
@@ -230,6 +263,48 @@ mod tests {
         };
     }
 
+    macro_rules! re {
+        () => {
+            None
+        };
+        ($re:expr) => {
+            Some(Regex::new($re).unwrap())
+        };
+    }
+
+    #[test]
+    fn asdf() {
+        let file_text = "foo bar baz";
+        let s = |a, b| SubMatch::new_text(a, b);
+        let test_cases = vec![
+            // no capture groups
+            (s("foo", 0..3), r"foo", re!(), r"foo bar baz"),
+            (s("foo", 0..3), r"bar", re!(), r"bar bar baz"),
+            (s("foo", 0..3), r"A", re!(), r"A bar baz"),
+            (s("foo", 0..3), r"Rust", re!(), r"Rust bar baz"),
+            (s("foo", 0..3), r"🦀", re!(), r"🦀 bar baz"),
+            (s("foo", 0..3), r"¯\_(ツ)_/¯", re!(), r"¯\_(ツ)_/¯ bar baz"),
+            // capture groups
+            (s("foo", 0..3), r"$1", re!("(foo)"), r"foo bar baz"),
+            (s("foo", 0..3), r"$1 $1", re!("(foo)"), r"foo foo bar baz"),
+            (s("foo", 0..3), r"bar$1", re!("(f)oo"), r"barf bar baz"),
+            (s("foo", 0..3), r"$2🦀$2", re!("(f)(o+)"), r"oo🦀oo bar baz"),
+            (
+                s("foo bar baz", 0..11),
+                r"¯\_(ツ)_/¯$3$4$1 ${2}f",
+                re!("(f)((o)(o)).*"),
+                r"¯\_(ツ)_/¯oof oof",
+            ),
+        ];
+
+        for (submatch, replacement, capture_pattern, expected) in test_cases {
+            let (item1, p1) = temp_item!(0, file_text, vec![submatch]);
+            let criteria = ReplacementCriteria::new(capture_pattern, replacement, vec![item1]);
+            perform_replacements(criteria).unwrap();
+            assert_eq!(fs::read_to_string(p1).unwrap(), expected);
+        }
+    }
+
     #[test]
     fn it_performs_replacements_only_on_match_items() {
         let text = "foo bar baz";
@@ -261,7 +336,7 @@ mod tests {
             build_item(RgMessageKind::Summary, &p5),
         ];
 
-        perform_replacements(ReplacementCriteria::new("NEW_VALUE", items)).unwrap();
+        perform_replacements(ReplacementCriteria::new(None, "NEW_VALUE", items)).unwrap();
         assert_eq!(fs::read_to_string(p1).unwrap(), text);
         assert_eq!(fs::read_to_string(p2).unwrap(), text);
         assert_eq!(fs::read_to_string(p3).unwrap(), "NEW_VALUE bar baz");
@@ -288,7 +363,7 @@ mod tests {
         assert_eq!(perms().mode(), 0o100777);
 
         // perform replacement
-        perform_replacements(ReplacementCriteria::new("NEW_VALUE", vec![item])).unwrap();
+        perform_replacements(ReplacementCriteria::new(None, "NEW_VALUE", vec![item])).unwrap();
         assert_eq!(fs::read_to_string(&path).unwrap(), "NEW_VALUE bar baz");
 
         // now check permissions are what we expect
@@ -302,7 +377,7 @@ mod tests {
         let (item3, p3) = temp_item!(0, "bar baz foo", vec![SubMatch::new_text("foo", 8..11)]);
 
         let items = vec![item1, item2, item3];
-        perform_replacements(ReplacementCriteria::new("NEW_VALUE", items)).unwrap();
+        perform_replacements(ReplacementCriteria::new(None, "NEW_VALUE", items)).unwrap();
         assert_eq!(fs::read_to_string(p1).unwrap(), "NEW_VALUE bar baz");
         assert_eq!(fs::read_to_string(p2).unwrap(), "baz NEW_VALUE bar");
         assert_eq!(fs::read_to_string(p3).unwrap(), "bar baz NEW_VALUE");
@@ -320,7 +395,7 @@ mod tests {
         items[1].set_should_replace(0, true);
         items[2].set_should_replace(0, false);
 
-        perform_replacements(ReplacementCriteria::new("NEW_VALUE", items)).unwrap();
+        perform_replacements(ReplacementCriteria::new(None, "NEW_VALUE", items)).unwrap();
         assert_eq!(fs::read_to_string(p1).unwrap(), "foo bar baz");
         assert_eq!(fs::read_to_string(p2).unwrap(), "baz NEW_VALUE bar");
         assert_eq!(fs::read_to_string(p3).unwrap(), "bar baz foo");
@@ -338,7 +413,7 @@ mod tests {
             ]
         );
 
-        perform_replacements(ReplacementCriteria::new("NEW_VALUE", vec![item])).unwrap();
+        perform_replacements(ReplacementCriteria::new(None, "NEW_VALUE", vec![item])).unwrap();
         assert_eq!(
             fs::read_to_string(p).unwrap(),
             "NEW_VALUE NEW_VALUE NEW_VALUE"
@@ -380,7 +455,7 @@ mod tests {
             ),
         ];
 
-        perform_replacements(ReplacementCriteria::new("NEW_VALUE", items)).unwrap();
+        perform_replacements(ReplacementCriteria::new(None, "NEW_VALUE", items)).unwrap();
         assert_eq!(
             fs::read_to_string(p).unwrap(),
             "NEW_VALUE bar baz\n...\nbaz NEW_VALUE bar\n...\nbar baz NEW_VALUE"
@@ -413,7 +488,7 @@ mod tests {
             ),
         ];
 
-        perform_replacements(ReplacementCriteria::new("NEW_VALUE", items)).unwrap();
+        perform_replacements(ReplacementCriteria::new(None, "NEW_VALUE", items)).unwrap();
         assert_eq!(
             fs::read_to_string(p).unwrap(),
             "foo bar baz\n...\nbaz NEW_VALUE bar\n...\nbar NEW_VALUE foo"
@@ -427,6 +502,7 @@ mod tests {
         use std::ffi::OsStr;
         use std::os::unix::ffi::OsStrExt;
         use std::path::PathBuf;
+
         use tempfile::tempdir;
 
         // Here, the values 0x66 and 0x6f correspond to 'f' and 'o'
@@ -450,7 +526,7 @@ mod tests {
                 .build(),
         );
 
-        perform_replacements(ReplacementCriteria::new(" on", vec![item])).unwrap();
+        perform_replacements(ReplacementCriteria::new(None, " on", vec![item])).unwrap();
         assert_eq!(fs::read_to_string(p).unwrap(), "hell on earth");
     }
 
@@ -479,7 +555,7 @@ mod tests {
                     })
                     .collect();
 
-                perform_replacements(ReplacementCriteria::new($replace, items)).unwrap();
+                perform_replacements(ReplacementCriteria::new(None, $replace, items)).unwrap();
 
                 // Read file bytes.
                 let mut file_bytes = vec![];
@@ -498,7 +574,6 @@ mod tests {
 
     // The following are generated with:
     //   printf "<BOM>%s" $(printf "foo bar baz\n...\nbaz foo bar\n...\nbar baz foo" | iconv -f UTF8 -t <ENCODING> | xxd -p -c 128)
-    // printf "efbbbf%s" $(printf "RUST bar baz\n...\nbaz RUST bar\n...\nbar baz RUST" | iconv -f UTF8 -t UTF8 | xxd -p -c 128)
 
     const UTF8_FOO: &str =
         "666f6f206261722062617a0a2e2e2e0a62617a20666f6f206261720a2e2e2e0a6261722062617a20666f6f";
@@ -524,67 +599,95 @@ mod tests {
     const UTF16BE_A: &str = "feff00410020006200610072002000620061007a000a002e002e002e000a00620061007a002000410020006200610072000a002e002e002e000a006200610072002000620061007a00200041";
     const UTF16LE_A: &str = "fffe410020006200610072002000620061007a000a002e002e002e000a00620061007a002000410020006200610072000a002e002e002e000a006200610072002000620061007a0020004100";
 
-    simple_test!(
-        multiline_longer_utf8,
-        UTF8_FOO,
-        UTF8_RUST,
+    // The following are generated with:
+    //   printf "<BOM>%s" $(printf "🦀 bar baz\n...\nbaz 🦀 bar\n...\nbar baz 🦀" | iconv -f UTF8 -t <ENCODING> | xxd -p -c 128)
+
+    const UTF8_EMOJI: &str = "f09fa680206261722062617a0a2e2e2e0a62617a20f09fa680206261720a2e2e2e0a6261722062617a20f09fa680";
+    const UTF8BOM_EMOJI: &str = "efbbbff09fa680206261722062617a0a2e2e2e0a62617a20f09fa680206261720a2e2e2e0a6261722062617a20f09fa680";
+    const UTF16BE_EMOJI: &str = "feffd83edd800020006200610072002000620061007a000a002e002e002e000a00620061007a0020d83edd800020006200610072000a002e002e002e000a006200610072002000620061007a0020d83edd80";
+    const UTF16LE_EMOJI: &str = "fffe3ed880dd20006200610072002000620061007a000a002e002e002e000a00620061007a0020003ed880dd20006200610072000a002e002e002e000a006200610072002000620061007a0020003ed880dd";
+
+    // The following are generated with:
+    //   printf "%s" $(printf "¯\_(ツ)_/¯ bar baz\n...\nbaz ¯\_(ツ)_/¯ bar\n...\nbar baz ¯\_(ツ)_/¯" | iconv -f UTF8 -t <ENCODING> | xxd -p -c 128)
+
+    const UTF8_UNICODE: &str = "c2af5c5f28e38384295f2fc2af206261722062617a0a2e2e2e0a62617a20c2af5c5f28e38384295f2fc2af206261720a2e2e2e0a6261722062617a20c2af5c5f28e38384295f2fc2af";
+    const UTF8BOM_UNICODE: &str = "efbbbfc2af5c5f28e38384295f2fc2af206261722062617a0a2e2e2e0a62617a20c2af5c5f28e38384295f2fc2af206261720a2e2e2e0a6261722062617a20c2af5c5f28e38384295f2fc2af";
+    const UTF16BE_UNICODE: &str = "feff00af005c005f002830c40029005f002f00af0020006200610072002000620061007a000a002e002e002e000a00620061007a002000af005c005f002830c40029005f002f00af0020006200610072000a002e002e002e000a006200610072002000620061007a002000af005c005f002830c40029005f002f00af";
+    const UTF16LE_UNICODE: &str = "fffeaf005c005f002800c43029005f002f00af0020006200610072002000620061007a000a002e002e002e000a00620061007a002000af005c005f002800c43029005f002f00af0020006200610072000a002e002e002e000a006200610072002000620061007a002000af005c005f002800c43029005f002f00af00";
+
+    macro_rules! simple_test_batch {
+        ($name:ident, $left:ident, $right:ident, $info:expr, $submatches:expr) => {
+            simple_test_batch!(@ [UTF8, UTF8BOM, UTF16BE, UTF16LE], $name, $left, $right, $info, $submatches);
+        };
+
+        (@ [$($enc:ident$(,)?)+], $name:ident, $left:ident, $right:ident, $info:expr, $submatches:expr) => {
+            paste::paste! {
+                $(
+                    simple_test!(
+                        [<multiline_ $name _ $enc:lower>],
+                        [<$enc _ $left:upper>],
+                        [<$enc _ $right:upper>],
+                        $info,
+                        // account for 3 byte BOM
+                        if stringify!($enc) == "UTF8BOM" {
+                            $submatches
+                                .into_iter()
+                                .map(|(a, b)| (a, std::ops::Range { start: b.start + 3, end: b.end + 3 }))
+                                .collect::<Vec<_>>()
+                        } else {
+                            $submatches
+                        }
+                    );
+                )+
+            }
+        };
+    }
+
+    simple_test_batch!(
+        to_longer,
+        FOO,
+        RUST,
         ("foo", "RUST"),
-        &[(0, 0..3), (16, 4..7), (32, 8..11)]
+        vec![(0, 0..3), (16, 4..7), (32, 8..11)]
     );
 
-    simple_test!(
-        multiline_longer_utf8_bom,
-        UTF8BOM_FOO,
-        UTF8BOM_RUST,
-        ("foo", "RUST"),
-        &[(0, 3..6), (19, 4..7), (35, 8..11)]
-    );
-
-    simple_test!(
-        multiline_longer_utf16be,
-        UTF16BE_FOO,
-        UTF16BE_RUST,
-        ("foo", "RUST"),
-        &[(0, 0..3), (16, 4..7), (32, 8..11)]
-    );
-
-    simple_test!(
-        multiline_longer_utf16le,
-        UTF16LE_FOO,
-        UTF16LE_RUST,
-        ("foo", "RUST"),
-        &[(0, 0..3), (16, 4..7), (32, 8..11)]
-    );
-
-    simple_test!(
-        multiline_shorter_utf8,
-        UTF8_FOO,
-        UTF8_A,
+    simple_test_batch!(
+        to_shorter,
+        FOO,
+        A,
         ("foo", "A"),
-        &[(0, 0..3), (16, 4..7), (32, 8..11)]
+        vec![(0, 0..3), (16, 4..7), (32, 8..11)]
     );
 
-    simple_test!(
-        multiline_shorter_utf8_bom,
-        UTF8BOM_FOO,
-        UTF8BOM_A,
-        ("foo", "A"),
-        &[(0, 3..6), (19, 4..7), (35, 8..11)]
+    simple_test_batch!(
+        to_emoji,
+        FOO,
+        EMOJI,
+        ("foo", "🦀"),
+        vec![(0, 0..3), (16, 4..7), (32, 8..11)]
     );
 
-    simple_test!(
-        multiline_shorter_utf16be,
-        UTF16BE_FOO,
-        UTF16BE_A,
-        ("foo", "A"),
-        &[(0, 0..3), (16, 4..7), (32, 8..11)]
+    simple_test_batch!(
+        from_emoji,
+        EMOJI,
+        FOO,
+        ("🦀", "foo"),
+        vec![(0, 0..4), (16, 5..9), (32, 10..14)]
     );
 
-    simple_test!(
-        multiline_shorter_utf16le,
-        UTF16LE_FOO,
-        UTF16LE_A,
-        ("foo", "A"),
-        &[(0, 0..3), (16, 4..7), (32, 8..11)]
+    simple_test_batch!(
+        to_unicode,
+        FOO,
+        UNICODE,
+        ("foo", r"¯\_(ツ)_/¯"),
+        vec![(0, 0..3), (16, 4..7), (32, 8..11)]
+    );
+
+    simple_test_batch!(
+        from_unicode,
+        UNICODE,
+        FOO,
+        (r"¯\_(ツ)_/¯", "foo"),
+        vec![(0, 0..13), (16, 14..27), (32, 28..41)]
     );
 }